backend/requirements.txt
Django==5.2.7
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
Pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
django-jazzmin==2.6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==24.3
